separando logica
